+++ simulator/examples/analog-clock.rs
@@ -13,7 +13,7 @@ use embedded_graphics::fonts::Font12x16;
 use embedded_graphics::pixelcolor::BinaryColor;
 use embedded_graphics::prelude::*;
 use embedded_graphics::primitives::{Circle, Line, Rectangle};
-use embedded_graphics_simulator::DisplayBuilder;
+use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
 use std::thread;
 use std::time::Duration;
 
@@ -133,11 +133,8 @@ fn draw_digital_clock<'a>(time_str: &'a str) -> impl Iterator<Item = Pixel<Binar
 }
 
 fn main() {
-    let mut display = DisplayBuilder::new()
-        .title("Clock")
-        .size(DISP_SIZE as usize, DISP_SIZE as usize)
-        .scale(2)
-        .build_binary();
+    let mut display = SimulatorDisplay::new(Size::new(DISP_SIZE as u32, DISP_SIZE as u32));
+    let mut window = WindowBuilder::new(&display).title("Clock").scale(2).build();
 
     loop {
         let time = Local::now();
@@ -151,7 +148,7 @@ fn main() {
             time.second()
         );
 
-        display.clear();
+        display.clear(BinaryColor::Off);
 
         draw_face().draw(&mut display);
         draw_hour_hand(time.hour()).draw(&mut display);
@@ -167,8 +164,9 @@ fn main() {
             .fill_color(Some(BinaryColor::On))
             .draw(&mut display);
 
-        let end = display.run_once();
+        window.update(&display);
 
+        let end = window.handle_events();
         if end {
             break;
         }