+++ simulator/examples/fonts.rs
@@ -2,15 +2,10 @@ use embedded_graphics::fonts::{Font6x12, Font6x8, Font8x16};
 use embedded_graphics::pixelcolor::BinaryColor;
 use embedded_graphics::prelude::*;
 use embedded_graphics::{text_12x16, text_6x8};
-use embedded_graphics_simulator::DisplayBuilder;
-use std::thread;
-use std::time::Duration;
+use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
 
 fn main() {
-    let mut display = DisplayBuilder::new()
-        .title("Fonts")
-        .size(256, 128)
-        .build_binary();
+    let mut display = SimulatorDisplay::new(Size::new(256, 128));
 
     // Show smallest font with black font on white background (default value for fonts)
     Font6x8::render_str("Hello World! - default style 6x8")
@@ -48,13 +43,6 @@ fn main() {
         .translate(Point::new(15, 105))
         .draw(&mut display);
 
-    loop {
-        let end = display.run_once();
-
-        if end {
-            break;
-        }
-
-        thread::sleep(Duration::from_millis(200));
-    }
+    let mut window = WindowBuilder::new(&display).title("Fonts").build();
+    window.show_static(&display);
 }