+++ simulator/examples/hello.rs
@@ -2,15 +2,10 @@ use embedded_graphics::fonts::Font6x8;
 use embedded_graphics::pixelcolor::BinaryColor;
 use embedded_graphics::prelude::*;
 use embedded_graphics::primitives::{Circle, Line};
-use embedded_graphics_simulator::{BinaryColorTheme, DisplayBuilder};
-use std::thread;
-use std::time::Duration;
+use embedded_graphics_simulator::{BinaryColorTheme, SimulatorDisplay, WindowBuilder};
 
 fn main() {
-    let mut display = DisplayBuilder::new()
-        .title("Hello World")
-        .theme(BinaryColorTheme::OledBlue)
-        .build_binary();
+    let mut display = SimulatorDisplay::new(Size::new(129, 129));
 
     // Outline
     Circle::new(Point::new(64, 64), 64)
@@ -30,13 +25,9 @@ fn main() {
         .translate(Point::new(5, 50))
         .draw(&mut display);
 
-    loop {
-        let end = display.run_once();
-
-        if end {
-            break;
-        }
-
-        thread::sleep(Duration::from_millis(200));
-    }
+    let mut window = WindowBuilder::new(&display)
+        .title("Hello World")
+        .theme(BinaryColorTheme::OledBlue)
+        .build();
+    window.show_static(&display);
 }