+++ simulator/examples/offscreen.rs
@@ -1,16 +1,10 @@
 use embedded_graphics::pixelcolor::BinaryColor;
 use embedded_graphics::prelude::*;
 use embedded_graphics::primitives::Rectangle;
-use embedded_graphics_simulator::DisplayBuilder;
-use std::thread;
-use std::time::Duration;
+use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
 
 fn main() {
-    let mut display = DisplayBuilder::new()
-        .title("Offscreen")
-        .size(32, 32)
-        .scale(4)
-        .build_binary();
+    let mut display = SimulatorDisplay::new(Size::new(32, 32));
 
     // Outline
     Rectangle::new(Point::new(0, 0), Point::new(16, 16))
@@ -18,13 +12,9 @@ fn main() {
         .translate(Point::new(-8, -8))
         .draw(&mut display);
 
-    loop {
-        let end = display.run_once();
-
-        if end {
-            break;
-        }
-
-        thread::sleep(Duration::from_millis(200));
-    }
+    let mut window = WindowBuilder::new(&display)
+        .title("Offscreen")
+        .scale(4)
+        .build();
+    window.show_static(&display);
 }