+++ simulator/examples/stroke.rs
@@ -1,17 +1,12 @@
 use embedded_graphics::pixelcolor::BinaryColor;
 use embedded_graphics::prelude::*;
 use embedded_graphics::primitives::{Circle, Line, Rectangle, Triangle};
-use embedded_graphics_simulator::DisplayBuilder;
-use std::thread;
-use std::time::Duration;
+use embedded_graphics_simulator::{SimulatorDisplay, WindowBuilder};
 
 const PADDING: i32 = 16;
 
 fn main() {
-    let mut display = DisplayBuilder::new()
-        .title("Strokes")
-        .size(320, 256)
-        .build_binary();
+    let mut display = SimulatorDisplay::new(Size::new(320, 256));
 
     let triangle = Triangle::new(Point::new(0, 64), Point::new(64, 0), Point::new(64, 64))
         .translate(Point::new(0, 0))
@@ -65,13 +60,6 @@ fn main() {
         )
         .draw(&mut display);
 
-    loop {
-        let end = display.run_once();
-
-        if end {
-            break;
-        }
-
-        thread::sleep(Duration::from_millis(200));
-    }
+    let mut window = WindowBuilder::new(&display).title("Strokes").build();
+    window.show_static(&display);
 }